requirements.txt
email-validator>=2.2.0
flask>=3.1.0
flask-sqlalchemy>=3.1.1
groq>=0.22.0
gunicorn>=23.0.0
openai>=1.71.0
psycopg2-binary>=2.9.10
pydub>=0.25.1
reportlab>=4.3.1
requests>=2.32.3
speechrecognition>=3.14.2
trafilatura>=2.0.0
youtube-transcript-api>=1.0.3
yt-dlp>=2025.3.31
lxml_html_clean
python-dotenv
